refactor(speech): extract cpal_io tests to sibling file
Move the single trailing `#[cfg(test)] mod tests` block out of
`primer-speech/src/cpal_io.rs` (742 lines) into a sibling
`src/cpal_io/tests.rs`. Pure verbatim relocation — no behaviour
change, no incidental production-code fixes.

`cpal_io.rs` is gated behind `#[cfg(feature = "cpal")]` and declared
in `lib.rs`, so its submodule resolves to `src/cpal_io/tests.rs`
(new `cpal_io/` dir). The tests are a descendant module, so
`use super::*` still reaches every private parent item — zero
visibility churn.

cpal_io.rs drops 742 -> 454 lines (-288, now under the ~500 line
guideline). The only parent-diff addition is `mod tests;`.

Verification:
- cargo test -p primer-speech --features cpal --lib: 45 passed,
  1 ignored (identical to baseline)
- cargo test -p primer-speech --lib (default): 33 passed
- cargo clippy -p primer-speech {--features cpal,} --all-targets
  -D warnings: clean
- cargo fmt --all --check: clean
- cargo test --workspace: 51 ok blocks, 0 failures
